worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Rosetta Celery Worker - Background task processing for heavy operations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.4.0,<6.0.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary==2.9.11",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv==1.0.0",
    "cryptography==42.0.0",
    "duckdb>=1.4.4",
    "adbc-driver-snowflake>=1.0.0",
    "pyarrow>=23.0.0",
    "structlog==24.1.0",
    "python-json-logger==2.0.7",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlglot>=26.0.0",
    "snowflake-connector-python>=3.0.0",
    "orjson>=3.10.0",
]

[dependency-groups]
dev = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "black==23.12.1",
    "flake8==7.0.0",
]

[tool.uv]
native-tls = true
